* Merge in IronPython 2.7 tests changes

* Fixes based on review

* Enable tests for IronLanguages/main#240

+++ Tests/interop/net/field/test_literal_fields.py
@@ -195,7 +195,7 @@ types = [
     GenericStructWithLiterals[int], 
     GenericStructWithLiterals[str], 
     ClassWithLiterals, 
-    GenericClassWithLiterals[long], 
+    GenericClassWithLiterals[int], 
     GenericClassWithLiterals[object],
     ]
 for i in range(len(types)):
